Network.AWS.DeviceFarm.CreateDevicePool

Description

Creates a device pool.

Creating a Request

createDevicePool

Arguments

:: Text
cdpProjectARN

-> Text
cdpName

-> CreateDevicePool

Creates a value of CreateDevicePool with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdpDescription - The device pool's description.

cdpProjectARN - The ARN of the project for the device pool.

cdpName - The device pool's name.

cdpRules - The device pool's rules.

data CreateDevicePool

Represents a request to the create device pool operation.

See: createDevicePool smart constructor.

Request Lenses

cdpDescription :: Lens' CreateDevicePool (Maybe Text)

The device pool's description.

cdpProjectARN :: Lens' CreateDevicePool Text

The ARN of the project for the device pool.

cdpName :: Lens' CreateDevicePool Text

The device pool's name.

cdpRules :: Lens' CreateDevicePool [Rule]

The device pool's rules.

Destructuring the Response

createDevicePoolResponse

Arguments

:: Int
cdprsResponseStatus

-> CreateDevicePoolResponse

Creates a value of CreateDevicePoolResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdprsDevicePool - The newly created device pool.

cdprsResponseStatus - -- | The response status code.

data CreateDevicePoolResponse

Represents the result of a create device pool request.

See: createDevicePoolResponse smart constructor.

Response Lenses

cdprsDevicePool :: Lens' CreateDevicePoolResponse (Maybe DevicePool)

The newly created device pool.

cdprsResponseStatus :: Lens' CreateDevicePoolResponse Int

- | The response status code.