Network.AWS.DeviceFarm.CreateNetworkProfile

Description

Creates a network profile.

Creating a Request

createNetworkProfile

Arguments

:: Text
cnpProjectARN

-> Text
cnpName

-> CreateNetworkProfile

Creates a value of CreateNetworkProfile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnpUplinkJitterMs - Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

cnpUplinkLossPercent - Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

cnpDownlinkJitterMs - Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

cnpDownlinkLossPercent - Proportion of received packets that fail to arrive from 0 to 100 percent.

cnpType - The type of network profile you wish to create. Valid values are listed below.

cnpUplinkDelayMs - Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

cnpUplinkBandwidthBits - The data throughput rate in bits per second, as an integer from 0 to 104857600.

cnpDescription - The description of the network profile.

cnpDownlinkDelayMs - Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

cnpDownlinkBandwidthBits - The data throughput rate in bits per second, as an integer from 0 to 104857600.

cnpProjectARN - The Amazon Resource Name (ARN) of the project for which you want to create a network profile.

cnpName - The name you wish to specify for the new network profile.

data CreateNetworkProfile

See: createNetworkProfile smart constructor.

Request Lenses

cnpUplinkJitterMs :: Lens' CreateNetworkProfile (Maybe Integer)

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

cnpUplinkLossPercent :: Lens' CreateNetworkProfile (Maybe Natural)

Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

cnpDownlinkJitterMs :: Lens' CreateNetworkProfile (Maybe Integer)

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

cnpDownlinkLossPercent :: Lens' CreateNetworkProfile (Maybe Natural)

Proportion of received packets that fail to arrive from 0 to 100 percent.

cnpType :: Lens' CreateNetworkProfile (Maybe NetworkProfileType)

The type of network profile you wish to create. Valid values are listed below.

cnpUplinkDelayMs :: Lens' CreateNetworkProfile (Maybe Integer)

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

cnpUplinkBandwidthBits :: Lens' CreateNetworkProfile (Maybe Integer)

The data throughput rate in bits per second, as an integer from 0 to 104857600.

cnpDescription :: Lens' CreateNetworkProfile (Maybe Text)

The description of the network profile.

cnpDownlinkDelayMs :: Lens' CreateNetworkProfile (Maybe Integer)

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

cnpDownlinkBandwidthBits :: Lens' CreateNetworkProfile (Maybe Integer)

The data throughput rate in bits per second, as an integer from 0 to 104857600.

cnpProjectARN :: Lens' CreateNetworkProfile Text

The Amazon Resource Name (ARN) of the project for which you want to create a network profile.

cnpName :: Lens' CreateNetworkProfile Text

The name you wish to specify for the new network profile.

Destructuring the Response

createNetworkProfileResponse

Arguments

:: Int
cnprsResponseStatus

-> CreateNetworkProfileResponse

Creates a value of CreateNetworkProfileResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnprsNetworkProfile - The network profile that is returned by the create network profile request.

cnprsResponseStatus - -- | The response status code.

data CreateNetworkProfileResponse

See: createNetworkProfileResponse smart constructor.

Response Lenses

cnprsNetworkProfile :: Lens' CreateNetworkProfileResponse (Maybe NetworkProfile)

The network profile that is returned by the create network profile request.

cnprsResponseStatus :: Lens' CreateNetworkProfileResponse Int

- | The response status code.