Network.AWS.DeviceFarm.CreateRemoteAccessSession

Description

Specifies and starts a remote access session.

Creating a Request

createRemoteAccessSession

Arguments

:: Text
crasProjectARN

-> Text
crasDeviceARN

-> CreateRemoteAccessSession

Creates a value of CreateRemoteAccessSession with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crasName - The name of the remote access session that you wish to create.

crasConfiguration - The configuration information for the remote access session request.

crasProjectARN - The Amazon Resource Name (ARN) of the project for which you want to create a remote access session.

crasDeviceARN - The Amazon Resource Name (ARN) of the device for which you want to create a remote access session.

data CreateRemoteAccessSession

Creates and submits a request to start a remote access session.

See: createRemoteAccessSession smart constructor.

Request Lenses

crasName :: Lens' CreateRemoteAccessSession (Maybe Text)

The name of the remote access session that you wish to create.

crasConfiguration :: Lens' CreateRemoteAccessSession (Maybe CreateRemoteAccessSessionConfiguration)

The configuration information for the remote access session request.

crasProjectARN :: Lens' CreateRemoteAccessSession Text

The Amazon Resource Name (ARN) of the project for which you want to create a remote access session.

crasDeviceARN :: Lens' CreateRemoteAccessSession Text

The Amazon Resource Name (ARN) of the device for which you want to create a remote access session.

Destructuring the Response

createRemoteAccessSessionResponse

Arguments

:: Int
crasrsResponseStatus

-> CreateRemoteAccessSessionResponse

Creates a value of CreateRemoteAccessSessionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crasrsRemoteAccessSession - A container that describes the remote access session when the request to create a remote access session is sent.

crasrsResponseStatus - -- | The response status code.

data CreateRemoteAccessSessionResponse

Represents the server response from a request to create a remote access session.

See: createRemoteAccessSessionResponse smart constructor.

Response Lenses

crasrsRemoteAccessSession :: Lens' CreateRemoteAccessSessionResponse (Maybe RemoteAccessSession)

A container that describes the remote access session when the request to create a remote access session is sent.

crasrsResponseStatus :: Lens' CreateRemoteAccessSessionResponse Int

- | The response status code.