Network.AWS.DeviceFarm.CreateUpload

Description

Uploads an app or test scripts.

Creating a Request

createUpload

Arguments

:: Text
cuProjectARN

-> Text
cuName

-> UploadType
cuType

-> CreateUpload

Creates a value of CreateUpload with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cuContentType - The upload's content type (for example, "application/octet-stream").

cuProjectARN - The ARN of the project for the upload.

cuName - The upload's file name. The name should not contain the / character. If uploading an iOS app, the file name needs to end with the .ipa extension. If uploading an Android app, the file name needs to end with the .apk extension. For all others, the file name must end with the .zip file extension.

cuType - The upload's upload type. Must be one of the following values: * ANDROID_APP: An Android upload. * IOS_APP: An iOS upload. * WEB_APP: A web appliction upload. * EXTERNAL_DATA: An external data upload. * APPIUM_JAVA_JUNIT_TEST_PACKAGE: An Appium Java JUnit test package upload. * APPIUM_JAVA_TESTNG_TEST_PACKAGE: An Appium Java TestNG test package upload. * APPIUM_PYTHON_TEST_PACKAGE: An Appium Python test package upload. * APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE: An Appium Java JUnit test package upload. * APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE: An Appium Java TestNG test package upload. * APPIUM_WEB_PYTHON_TEST_PACKAGE: An Appium Python test package upload. * CALABASH_TEST_PACKAGE: A Calabash test package upload. * INSTRUMENTATION_TEST_PACKAGE: An instrumentation upload. * UIAUTOMATION_TEST_PACKAGE: A uiautomation test package upload. * UIAUTOMATOR_TEST_PACKAGE: A uiautomator test package upload. * XCTEST_TEST_PACKAGE: An XCode test package upload. * XCTEST_UI_TEST_PACKAGE: An XCode UI test package upload. Note If you call CreateUpload with WEB_APP specified, AWS Device Farm throws an ArgumentException error.

data CreateUpload

Represents a request to the create upload operation.

See: createUpload smart constructor.

Request Lenses

cuContentType :: Lens' CreateUpload (Maybe Text)

The upload's content type (for example, "application/octet-stream").

cuProjectARN :: Lens' CreateUpload Text

The ARN of the project for the upload.

cuName :: Lens' CreateUpload Text

The upload's file name. The name should not contain the / character. If uploading an iOS app, the file name needs to end with the .ipa extension. If uploading an Android app, the file name needs to end with the .apk extension. For all others, the file name must end with the .zip file extension.

cuType :: Lens' CreateUpload UploadType

The upload's upload type. Must be one of the following values: * ANDROID_APP: An Android upload. * IOS_APP: An iOS upload. * WEB_APP: A web appliction upload. * EXTERNAL_DATA: An external data upload. * APPIUM_JAVA_JUNIT_TEST_PACKAGE: An Appium Java JUnit test package upload. * APPIUM_JAVA_TESTNG_TEST_PACKAGE: An Appium Java TestNG test package upload. * APPIUM_PYTHON_TEST_PACKAGE: An Appium Python test package upload. * APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE: An Appium Java JUnit test package upload. * APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE: An Appium Java TestNG test package upload. * APPIUM_WEB_PYTHON_TEST_PACKAGE: An Appium Python test package upload. * CALABASH_TEST_PACKAGE: A Calabash test package upload. * INSTRUMENTATION_TEST_PACKAGE: An instrumentation upload. * UIAUTOMATION_TEST_PACKAGE: A uiautomation test package upload. * UIAUTOMATOR_TEST_PACKAGE: A uiautomator test package upload. * XCTEST_TEST_PACKAGE: An XCode test package upload. * XCTEST_UI_TEST_PACKAGE: An XCode UI test package upload. Note If you call CreateUpload with WEB_APP specified, AWS Device Farm throws an ArgumentException error.

Destructuring the Response

createUploadResponse

Arguments

:: Int
cursResponseStatus

-> CreateUploadResponse

Creates a value of CreateUploadResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cursUpload - The newly created upload.

cursResponseStatus - -- | The response status code.

data CreateUploadResponse

Represents the result of a create upload request.

See: createUploadResponse smart constructor.

Response Lenses

cursUpload :: Lens' CreateUploadResponse (Maybe Upload)

The newly created upload.

cursResponseStatus :: Lens' CreateUploadResponse Int

- | The response status code.