Network.AWS.DeviceFarm.DeleteProject

Description

Deletes an AWS Device Farm project, given the project ARN.

Note Deleting this resource does not stop an in-progress run.

Creating a Request

deleteProject

Arguments

:: Text
dpArn

-> DeleteProject

Creates a value of DeleteProject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpArn - Represents the Amazon Resource Name (ARN) of the Device Farm project you wish to delete.

data DeleteProject

Represents a request to the delete project operation.

See: deleteProject smart constructor.

Request Lenses

dpArn :: Lens' DeleteProject Text

Represents the Amazon Resource Name (ARN) of the Device Farm project you wish to delete.

Destructuring the Response

deleteProjectResponse

Arguments

:: Int
dprsResponseStatus

-> DeleteProjectResponse

Creates a value of DeleteProjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dprsResponseStatus - -- | The response status code.

data DeleteProjectResponse

Represents the result of a delete project request.

See: deleteProjectResponse smart constructor.

Response Lenses

dprsResponseStatus :: Lens' DeleteProjectResponse Int

- | The response status code.