Network.AWS.DeviceFarm.GetDevicePoolCompatibility

Description

Gets information about compatibility with a device pool.

Creating a Request

getDevicePoolCompatibility

Arguments

:: Text
gdpcDevicePoolARN

-> GetDevicePoolCompatibility

Creates a value of GetDevicePoolCompatibility with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdpcTest - Information about the uploaded test to be run against the device pool.

gdpcAppARN - The ARN of the app that is associated with the specified device pool.

gdpcTestType - The test type for the specified device pool. Allowed values include the following: * BUILTIN_FUZZ: The built-in fuzz type. * BUILTIN_EXPLORER: For Android, an app explorer that will traverse an Android app, interacting with it and capturing screenshots at the same time. * APPIUM_JAVA_JUNIT: The Appium Java JUnit type. * APPIUM_JAVA_TESTNG: The Appium Java TestNG type. * APPIUM_PYTHON: The Appium Python type. * APPIUM_WEB_JAVA_JUNIT: The Appium Java JUnit type for Web apps. * APPIUM_WEB_JAVA_TESTNG: The Appium Java TestNG type for Web apps. * APPIUM_WEB_PYTHON: The Appium Python type for Web apps. * CALABASH: The Calabash type. * INSTRUMENTATION: The Instrumentation type. * UIAUTOMATION: The uiautomation type. * UIAUTOMATOR: The uiautomator type. * XCTEST: The XCode test type. * XCTEST_UI: The XCode UI test type.

gdpcDevicePoolARN - The device pool's ARN.

data GetDevicePoolCompatibility

Represents a request to the get device pool compatibility operation.

See: getDevicePoolCompatibility smart constructor.

Request Lenses

gdpcTest :: Lens' GetDevicePoolCompatibility (Maybe ScheduleRunTest)

Information about the uploaded test to be run against the device pool.

gdpcAppARN :: Lens' GetDevicePoolCompatibility (Maybe Text)

The ARN of the app that is associated with the specified device pool.

gdpcTestType :: Lens' GetDevicePoolCompatibility (Maybe TestType)

The test type for the specified device pool. Allowed values include the following: * BUILTIN_FUZZ: The built-in fuzz type. * BUILTIN_EXPLORER: For Android, an app explorer that will traverse an Android app, interacting with it and capturing screenshots at the same time. * APPIUM_JAVA_JUNIT: The Appium Java JUnit type. * APPIUM_JAVA_TESTNG: The Appium Java TestNG type. * APPIUM_PYTHON: The Appium Python type. * APPIUM_WEB_JAVA_JUNIT: The Appium Java JUnit type for Web apps. * APPIUM_WEB_JAVA_TESTNG: The Appium Java TestNG type for Web apps. * APPIUM_WEB_PYTHON: The Appium Python type for Web apps. * CALABASH: The Calabash type. * INSTRUMENTATION: The Instrumentation type. * UIAUTOMATION: The uiautomation type. * UIAUTOMATOR: The uiautomator type. * XCTEST: The XCode test type. * XCTEST_UI: The XCode UI test type.

gdpcDevicePoolARN :: Lens' GetDevicePoolCompatibility Text

The device pool's ARN.

Destructuring the Response

getDevicePoolCompatibilityResponse

Arguments

:: Int
gdpcrsResponseStatus

-> GetDevicePoolCompatibilityResponse

Creates a value of GetDevicePoolCompatibilityResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdpcrsIncompatibleDevices - Information about incompatible devices.

gdpcrsCompatibleDevices - Information about compatible devices.

gdpcrsResponseStatus - -- | The response status code.

data GetDevicePoolCompatibilityResponse

Represents the result of describe device pool compatibility request.

See: getDevicePoolCompatibilityResponse smart constructor.

Response Lenses

gdpcrsIncompatibleDevices :: Lens' GetDevicePoolCompatibilityResponse [DevicePoolCompatibilityResult]

Information about incompatible devices.

gdpcrsCompatibleDevices :: Lens' GetDevicePoolCompatibilityResponse [DevicePoolCompatibilityResult]

Information about compatible devices.

gdpcrsResponseStatus :: Lens' GetDevicePoolCompatibilityResponse Int

- | The response status code.