Network.AWS.DeviceFarm.GetProject

Description

Gets information about a project.

Creating a Request

getProject

Arguments

:: Text
gpArn

-> GetProject

Creates a value of GetProject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpArn - The project's ARN.

data GetProject

Represents a request to the get project operation.

See: getProject smart constructor.

Request Lenses

gpArn :: Lens' GetProject Text

The project's ARN.

Destructuring the Response

getProjectResponse

Arguments

:: Int
gprsResponseStatus

-> GetProjectResponse

Creates a value of GetProjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gprsProject - The project you wish to get information about.

gprsResponseStatus - -- | The response status code.

data GetProjectResponse

Represents the result of a get project request.

See: getProjectResponse smart constructor.

Response Lenses

gprsProject :: Lens' GetProjectResponse (Maybe Project)

The project you wish to get information about.

gprsResponseStatus :: Lens' GetProjectResponse Int

- | The response status code.