Network.AWS.DeviceFarm.GetRun

Description

Gets information about a run.

Creating a Request

getRun

Arguments

:: Text
grArn

-> GetRun

Creates a value of GetRun with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grArn - The run's ARN.

data GetRun

Represents a request to the get run operation.

See: getRun smart constructor.

Request Lenses

grArn :: Lens' GetRun Text

The run's ARN.

Destructuring the Response

getRunResponse

Arguments

:: Int
grrsResponseStatus

-> GetRunResponse

Creates a value of GetRunResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grrsRun - The run you wish to get results from.

grrsResponseStatus - -- | The response status code.

data GetRunResponse

Represents the result of a get run request.

See: getRunResponse smart constructor.

Response Lenses

grrsRun :: Lens' GetRunResponse (Maybe Run)

The run you wish to get results from.

grrsResponseStatus :: Lens' GetRunResponse Int

- | The response status code.