Network.AWS.DeviceFarm.ListNetworkProfiles

Description

Returns the list of available network profiles.

Creating a Request

listNetworkProfiles

Arguments

:: Text
lnpArn

-> ListNetworkProfiles

Creates a value of ListNetworkProfiles with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lnpNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lnpType - The type of network profile you wish to return information about. Valid values are listed below.

lnpArn - The Amazon Resource Name (ARN) of the project for which you want to list network profiles.

data ListNetworkProfiles

See: listNetworkProfiles smart constructor.

Request Lenses

lnpNextToken :: Lens' ListNetworkProfiles (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lnpType :: Lens' ListNetworkProfiles (Maybe NetworkProfileType)

The type of network profile you wish to return information about. Valid values are listed below.

lnpArn :: Lens' ListNetworkProfiles Text

The Amazon Resource Name (ARN) of the project for which you want to list network profiles.

Destructuring the Response

listNetworkProfilesResponse

Arguments

:: Int
lnprsResponseStatus

-> ListNetworkProfilesResponse

Creates a value of ListNetworkProfilesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lnprsNetworkProfiles - A list of the available network profiles.

lnprsNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lnprsResponseStatus - -- | The response status code.

data ListNetworkProfilesResponse

See: listNetworkProfilesResponse smart constructor.

Response Lenses

lnprsNetworkProfiles :: Lens' ListNetworkProfilesResponse [NetworkProfile]

A list of the available network profiles.

lnprsNextToken :: Lens' ListNetworkProfilesResponse (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lnprsResponseStatus :: Lens' ListNetworkProfilesResponse Int

- | The response status code.