Network.AWS.DeviceFarm.ListOfferingPromotions

Description

Returns a list of offering promotions. Each offering promotion record contains the ID and description of the promotion. The API returns a NotEligible error if the caller is not permitted to invoke the operation. Contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

Creating a Request

listOfferingPromotions :: ListOfferingPromotions

Creates a value of ListOfferingPromotions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lopNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

data ListOfferingPromotions

See: listOfferingPromotions smart constructor.

Request Lenses

lopNextToken :: Lens' ListOfferingPromotions (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Destructuring the Response

listOfferingPromotionsResponse

Arguments

:: Int
loprsResponseStatus

-> ListOfferingPromotionsResponse

Creates a value of ListOfferingPromotionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

loprsNextToken - An identifier to be used in the next call to this operation, to return the next set of items in the list.

loprsOfferingPromotions - Information about the offering promotions.

loprsResponseStatus - -- | The response status code.

data ListOfferingPromotionsResponse

See: listOfferingPromotionsResponse smart constructor.

Response Lenses

loprsNextToken :: Lens' ListOfferingPromotionsResponse (Maybe Text)

An identifier to be used in the next call to this operation, to return the next set of items in the list.

loprsOfferingPromotions :: Lens' ListOfferingPromotionsResponse [OfferingPromotion]

Information about the offering promotions.

loprsResponseStatus :: Lens' ListOfferingPromotionsResponse Int

- | The response status code.