Network.AWS.DeviceFarm.ListOfferingTransactions

Description

Returns a list of all historical purchases, renewals, and system renewal transactions for an AWS account. The list is paginated and ordered by a descending timestamp (most recent transactions are first). The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

This operation returns paginated results.

Creating a Request

listOfferingTransactions :: ListOfferingTransactions

Creates a value of ListOfferingTransactions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lotNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

data ListOfferingTransactions

Represents the request to list the offering transaction history.

See: listOfferingTransactions smart constructor.

Request Lenses

lotNextToken :: Lens' ListOfferingTransactions (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Destructuring the Response

listOfferingTransactionsResponse

Arguments

:: Int
lotrsResponseStatus

-> ListOfferingTransactionsResponse

Creates a value of ListOfferingTransactionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lotrsOfferingTransactions - The audit log of subscriptions you have purchased and modified through AWS Device Farm.

lotrsNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lotrsResponseStatus - -- | The response status code.

data ListOfferingTransactionsResponse

Returns the transaction log of the specified offerings.

See: listOfferingTransactionsResponse smart constructor.

Response Lenses

lotrsOfferingTransactions :: Lens' ListOfferingTransactionsResponse [OfferingTransaction]

The audit log of subscriptions you have purchased and modified through AWS Device Farm.

lotrsNextToken :: Lens' ListOfferingTransactionsResponse (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lotrsResponseStatus :: Lens' ListOfferingTransactionsResponse Int

- | The response status code.