Network.AWS.DeviceFarm.ListOfferings

Description

Returns a list of products or offerings that the user can manage through the API. Each offering record indicates the recurring price per unit and the frequency for that offering. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

This operation returns paginated results.

Creating a Request

listOfferings :: ListOfferings

Creates a value of ListOfferings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

loNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

data ListOfferings

Represents the request to list all offerings.

See: listOfferings smart constructor.

Request Lenses

loNextToken :: Lens' ListOfferings (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Destructuring the Response

listOfferingsResponse

Arguments

:: Int
lorsResponseStatus

-> ListOfferingsResponse

Creates a value of ListOfferingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lorsNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lorsOfferings - A value representing the list offering results.

lorsResponseStatus - -- | The response status code.

data ListOfferingsResponse

Represents the return values of the list of offerings.

See: listOfferingsResponse smart constructor.

Response Lenses

lorsNextToken :: Lens' ListOfferingsResponse (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lorsOfferings :: Lens' ListOfferingsResponse [Offering]

A value representing the list offering results.

lorsResponseStatus :: Lens' ListOfferingsResponse Int

- | The response status code.