Network.AWS.DeviceFarm.ListProjects

Description

Gets information about projects.

This operation returns paginated results.

Creating a Request

listProjects :: ListProjects

Creates a value of ListProjects with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpArn - Optional. If no Amazon Resource Name (ARN) is specified, then AWS Device Farm returns a list of all projects for the AWS account. You can also specify a project ARN.

lpNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

data ListProjects

Represents a request to the list projects operation.

See: listProjects smart constructor.

Request Lenses

lpArn :: Lens' ListProjects (Maybe Text)

Optional. If no Amazon Resource Name (ARN) is specified, then AWS Device Farm returns a list of all projects for the AWS account. You can also specify a project ARN.

lpNextToken :: Lens' ListProjects (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Destructuring the Response

listProjectsResponse

Arguments

:: Int
lprsResponseStatus

-> ListProjectsResponse

Creates a value of ListProjectsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lprsNextToken - If the number of items that are returned is significantly large, this is an identifier that is also returned, which can be used in a subsequent call to this operation to return the next set of items in the list.

lprsProjects - Information about the projects.

lprsResponseStatus - -- | The response status code.

data ListProjectsResponse

Represents the result of a list projects request.

See: listProjectsResponse smart constructor.

Response Lenses

lprsNextToken :: Lens' ListProjectsResponse (Maybe Text)

If the number of items that are returned is significantly large, this is an identifier that is also returned, which can be used in a subsequent call to this operation to return the next set of items in the list.

lprsProjects :: Lens' ListProjectsResponse [Project]

Information about the projects.

lprsResponseStatus :: Lens' ListProjectsResponse Int

- | The response status code.