Network.AWS.DeviceFarm.ListSamples

Description

Gets information about samples, given an AWS Device Farm project ARN

This operation returns paginated results.

Creating a Request

listSamples

Arguments

:: Text
lsArn

-> ListSamples

Creates a value of ListSamples with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lsArn - The Amazon Resource Name (ARN) of the project for which you want to list samples.

data ListSamples

Represents a request to the list samples operation.

See: listSamples smart constructor.

Request Lenses

lsNextToken :: Lens' ListSamples (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lsArn :: Lens' ListSamples Text

The Amazon Resource Name (ARN) of the project for which you want to list samples.

Destructuring the Response

listSamplesResponse

Arguments

:: Int
lrsResponseStatus

-> ListSamplesResponse

Creates a value of ListSamplesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrsNextToken - If the number of items that are returned is significantly large, this is an identifier that is also returned, which can be used in a subsequent call to this operation to return the next set of items in the list.

lrsSamples - Information about the samples.

lrsResponseStatus - -- | The response status code.

data ListSamplesResponse

Represents the result of a list samples request.

See: listSamplesResponse smart constructor.

Response Lenses

lrsNextToken :: Lens' ListSamplesResponse (Maybe Text)

If the number of items that are returned is significantly large, this is an identifier that is also returned, which can be used in a subsequent call to this operation to return the next set of items in the list.

lrsSamples :: Lens' ListSamplesResponse [Sample]

Information about the samples.

lrsResponseStatus :: Lens' ListSamplesResponse Int

- | The response status code.