Network.AWS.DeviceFarm.ListSuites

Description

Gets information about suites.

This operation returns paginated results.

Creating a Request

listSuites

Arguments

:: Text
lArn

-> ListSuites

Creates a value of ListSuites with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lArn - The suites' ARNs.

data ListSuites

Represents a request to the list suites operation.

See: listSuites smart constructor.

Request Lenses

lNextToken :: Lens' ListSuites (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lArn :: Lens' ListSuites Text

The suites' ARNs.

Destructuring the Response

listSuitesResponse

Arguments

:: Int
lsrsResponseStatus

-> ListSuitesResponse

Creates a value of ListSuitesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsrsNextToken - If the number of items that are returned is significantly large, this is an identifier that is also returned, which can be used in a subsequent call to this operation to return the next set of items in the list.

lsrsSuites - Information about the suites.

lsrsResponseStatus - -- | The response status code.

data ListSuitesResponse

Represents the result of a list suites request.

See: listSuitesResponse smart constructor.

Response Lenses

lsrsNextToken :: Lens' ListSuitesResponse (Maybe Text)

If the number of items that are returned is significantly large, this is an identifier that is also returned, which can be used in a subsequent call to this operation to return the next set of items in the list.

lsrsSuites :: Lens' ListSuitesResponse [Suite]

Information about the suites.

lsrsResponseStatus :: Lens' ListSuitesResponse Int

- | The response status code.