Network.AWS.DeviceFarm.ListTests

Description

Gets information about tests.

This operation returns paginated results.

Creating a Request

listTests

Arguments

:: Text
ltArn

-> ListTests

Creates a value of ListTests with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

ltArn - The tests' ARNs.

data ListTests

Represents a request to the list tests operation.

See: listTests smart constructor.

Request Lenses

ltNextToken :: Lens' ListTests (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

ltArn :: Lens' ListTests Text

The tests' ARNs.

Destructuring the Response

listTestsResponse

Arguments

:: Int
ltrsResponseStatus

-> ListTestsResponse

Creates a value of ListTestsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltrsTests - Information about the tests.

ltrsNextToken - If the number of items that are returned is significantly large, this is an identifier that is also returned, which can be used in a subsequent call to this operation to return the next set of items in the list.

ltrsResponseStatus - -- | The response status code.

data ListTestsResponse

Represents the result of a list tests request.

See: listTestsResponse smart constructor.

Response Lenses

ltrsTests :: Lens' ListTestsResponse [Test]

Information about the tests.

ltrsNextToken :: Lens' ListTestsResponse (Maybe Text)

If the number of items that are returned is significantly large, this is an identifier that is also returned, which can be used in a subsequent call to this operation to return the next set of items in the list.

ltrsResponseStatus :: Lens' ListTestsResponse Int

- | The response status code.