Network.AWS.DeviceFarm.ListUploads

Description

Gets information about uploads, given an AWS Device Farm project ARN.

This operation returns paginated results.

Creating a Request

listUploads

Arguments

:: Text
luArn

-> ListUploads

Creates a value of ListUploads with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

luNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

luArn - The Amazon Resource Name (ARN) of the project for which you want to list uploads.

data ListUploads

Represents a request to the list uploads operation.

See: listUploads smart constructor.

Request Lenses

luNextToken :: Lens' ListUploads (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

luArn :: Lens' ListUploads Text

The Amazon Resource Name (ARN) of the project for which you want to list uploads.

Destructuring the Response

listUploadsResponse

Arguments

:: Int
lursResponseStatus

-> ListUploadsResponse

Creates a value of ListUploadsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lursNextToken - If the number of items that are returned is significantly large, this is an identifier that is also returned, which can be used in a subsequent call to this operation to return the next set of items in the list.

lursUploads - Information about the uploads.

lursResponseStatus - -- | The response status code.

data ListUploadsResponse

Represents the result of a list uploads request.

See: listUploadsResponse smart constructor.

Response Lenses

lursNextToken :: Lens' ListUploadsResponse (Maybe Text)

If the number of items that are returned is significantly large, this is an identifier that is also returned, which can be used in a subsequent call to this operation to return the next set of items in the list.

lursUploads :: Lens' ListUploadsResponse [Upload]

Information about the uploads.

lursResponseStatus :: Lens' ListUploadsResponse Int

- | The response status code.