Network.AWS.DeviceFarm.PurchaseOffering

Description

Immediately purchases offerings for an AWS account. Offerings renew with the latest total purchased quantity for an offering, unless the renewal was overridden. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

Creating a Request

purchaseOffering :: PurchaseOffering

Creates a value of PurchaseOffering with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

poQuantity - The number of device slots you wish to purchase in an offering request.

poOfferingId - The ID of the offering.

poOfferingPromotionId - The ID of the offering promotion to be applied to the purchase.

data PurchaseOffering

Represents a request for a purchase offering.

See: purchaseOffering smart constructor.

Request Lenses

poQuantity :: Lens' PurchaseOffering (Maybe Int)

The number of device slots you wish to purchase in an offering request.

poOfferingId :: Lens' PurchaseOffering (Maybe Text)

The ID of the offering.

poOfferingPromotionId :: Lens' PurchaseOffering (Maybe Text)

The ID of the offering promotion to be applied to the purchase.

Destructuring the Response

purchaseOfferingResponse

Arguments

:: Int
porsResponseStatus

-> PurchaseOfferingResponse

Creates a value of PurchaseOfferingResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

porsOfferingTransaction - Represents the offering transaction for the purchase result.

porsResponseStatus - -- | The response status code.

data PurchaseOfferingResponse

The result of the purchase offering (e.g., success or failure).

See: purchaseOfferingResponse smart constructor.

Response Lenses

porsOfferingTransaction :: Lens' PurchaseOfferingResponse (Maybe OfferingTransaction)

Represents the offering transaction for the purchase result.

porsResponseStatus :: Lens' PurchaseOfferingResponse Int

- | The response status code.