Network.AWS.DeviceFarm.StopRemoteAccessSession

Description

Ends a specified remote access session.

Creating a Request

stopRemoteAccessSession

Arguments

:: Text
srasArn

-> StopRemoteAccessSession

Creates a value of StopRemoteAccessSession with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srasArn - The Amazon Resource Name (ARN) of the remote access session you wish to stop.

data StopRemoteAccessSession

Represents the request to stop the remote access session.

See: stopRemoteAccessSession smart constructor.

Request Lenses

srasArn :: Lens' StopRemoteAccessSession Text

The Amazon Resource Name (ARN) of the remote access session you wish to stop.

Destructuring the Response

stopRemoteAccessSessionResponse

Arguments

:: Int
srasrsResponseStatus

-> StopRemoteAccessSessionResponse

Creates a value of StopRemoteAccessSessionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srasrsRemoteAccessSession - A container representing the metadata from the service about the remote access session you are stopping.

srasrsResponseStatus - -- | The response status code.

data StopRemoteAccessSessionResponse

Represents the response from the server that describes the remote access session when AWS Device Farm stops the session.

See: stopRemoteAccessSessionResponse smart constructor.

Response Lenses

srasrsRemoteAccessSession :: Lens' StopRemoteAccessSessionResponse (Maybe RemoteAccessSession)

A container representing the metadata from the service about the remote access session you are stopping.

srasrsResponseStatus :: Lens' StopRemoteAccessSessionResponse Int

- | The response status code.