Network.AWS.DeviceFarm.UpdateDevicePool

Description

Modifies the name, description, and rules in a device pool given the attributes and the pool ARN. Rule updates are all-or-nothing, meaning they can only be updated as a whole (or not at all).

Creating a Request

updateDevicePool

Arguments

:: Text
udpArn

-> UpdateDevicePool

Creates a value of UpdateDevicePool with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udpRules - Represents the rules you wish to modify for the device pool. Updating rules is optional; however, if you choose to update rules for your request, the update will replace the existing rules.

udpName - A string representing the name of the device pool you wish to update.

udpDescription - A description of the device pool you wish to update.

udpArn - The Amazon Resourc Name (ARN) of the Device Farm device pool you wish to update.

data UpdateDevicePool

Represents a request to the update device pool operation.

See: updateDevicePool smart constructor.

Request Lenses

udpRules :: Lens' UpdateDevicePool [Rule]

Represents the rules you wish to modify for the device pool. Updating rules is optional; however, if you choose to update rules for your request, the update will replace the existing rules.

udpName :: Lens' UpdateDevicePool (Maybe Text)

A string representing the name of the device pool you wish to update.

udpDescription :: Lens' UpdateDevicePool (Maybe Text)

A description of the device pool you wish to update.

udpArn :: Lens' UpdateDevicePool Text

The Amazon Resourc Name (ARN) of the Device Farm device pool you wish to update.

Destructuring the Response

updateDevicePoolResponse

Arguments

:: Int
udprsResponseStatus

-> UpdateDevicePoolResponse

Creates a value of UpdateDevicePoolResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udprsDevicePool - The device pool you just updated.

udprsResponseStatus - -- | The response status code.

data UpdateDevicePoolResponse

Represents the result of an update device pool request.

See: updateDevicePoolResponse smart constructor.

Response Lenses

udprsDevicePool :: Lens' UpdateDevicePoolResponse (Maybe DevicePool)

The device pool you just updated.

udprsResponseStatus :: Lens' UpdateDevicePoolResponse Int

- | The response status code.