Network.AWS.DeviceFarm.UpdateNetworkProfile

Description

Updates the network profile with specific settings.

Creating a Request

updateNetworkProfile

Arguments

:: Text
unpArn

-> UpdateNetworkProfile

Creates a value of UpdateNetworkProfile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

unpUplinkJitterMs - Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

unpUplinkLossPercent - Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

unpDownlinkJitterMs - Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

unpName - The name of the network profile about which you are returning information.

unpDownlinkLossPercent - Proportion of received packets that fail to arrive from 0 to 100 percent.

unpType - The type of network profile you wish to return information about. Valid values are listed below.

unpUplinkDelayMs - Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

unpUplinkBandwidthBits - The data throughput rate in bits per second, as an integer from 0 to 104857600.

unpDescription - The descriptoin of the network profile about which you are returning information.

unpDownlinkDelayMs - Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

unpDownlinkBandwidthBits - The data throughput rate in bits per second, as an integer from 0 to 104857600.

unpArn - The Amazon Resource Name (ARN) of the project that you wish to update network profile settings.

data UpdateNetworkProfile

See: updateNetworkProfile smart constructor.

Request Lenses

unpUplinkJitterMs :: Lens' UpdateNetworkProfile (Maybe Integer)

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

unpUplinkLossPercent :: Lens' UpdateNetworkProfile (Maybe Natural)

Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

unpDownlinkJitterMs :: Lens' UpdateNetworkProfile (Maybe Integer)

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

unpName :: Lens' UpdateNetworkProfile (Maybe Text)

The name of the network profile about which you are returning information.

unpDownlinkLossPercent :: Lens' UpdateNetworkProfile (Maybe Natural)

Proportion of received packets that fail to arrive from 0 to 100 percent.

unpType :: Lens' UpdateNetworkProfile (Maybe NetworkProfileType)

The type of network profile you wish to return information about. Valid values are listed below.

unpUplinkDelayMs :: Lens' UpdateNetworkProfile (Maybe Integer)

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

unpUplinkBandwidthBits :: Lens' UpdateNetworkProfile (Maybe Integer)

The data throughput rate in bits per second, as an integer from 0 to 104857600.

unpDescription :: Lens' UpdateNetworkProfile (Maybe Text)

The descriptoin of the network profile about which you are returning information.

unpDownlinkDelayMs :: Lens' UpdateNetworkProfile (Maybe Integer)

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

unpDownlinkBandwidthBits :: Lens' UpdateNetworkProfile (Maybe Integer)

The data throughput rate in bits per second, as an integer from 0 to 104857600.

unpArn :: Lens' UpdateNetworkProfile Text

The Amazon Resource Name (ARN) of the project that you wish to update network profile settings.

Destructuring the Response

updateNetworkProfileResponse

Arguments

:: Int
unprsResponseStatus

-> UpdateNetworkProfileResponse

Creates a value of UpdateNetworkProfileResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

unprsNetworkProfile - A list of the available network profiles.

unprsResponseStatus - -- | The response status code.

data UpdateNetworkProfileResponse

See: updateNetworkProfileResponse smart constructor.

Response Lenses

unprsNetworkProfile :: Lens' UpdateNetworkProfileResponse (Maybe NetworkProfile)

A list of the available network profiles.

unprsResponseStatus :: Lens' UpdateNetworkProfileResponse Int

- | The response status code.